requirements.txt
pandas==1.4.3
pywin32==304
PyYAML==6.0
